app/requirements.txt
Django==2.0.3
djangorestframework==3.8.2
drf-extensions==0.3.1
pytz==2018.3
